refactor(questions): namespace under Rounds and drop dead scaffold
Only create + destroy were ever used; index/show/new/edit/update were scaffold
leftovers (no templates, not linked anywhere — question text is saved via the
round form's nested questions_attributes, not questions#update).

- move the controller to Rounds::QuestionsController (app/controllers/rounds/),
  keeping only create (nested) and destroy (shallow)
- routes: questions create nested under rounds, destroy shallow, both via
  module: :rounds — path helpers (round_questions_path, question_path) unchanged
- delete the flat QuestionsController, the unused questions/_form & _question
  views, and the dead scaffold + system tests
- move controller tests to test/controllers/rounds/, trimmed to the live actions
  plus an auth check
