--- app/views/questions/_question.html.slim
@@ -1,9 +1,0 @@
-p
-  strong Question:
-  = question.question
-
-  strong Answer:
-  = question.answer
-
-= link_to 'Edit', edit_question_path(question)
-= link_to 'Back', questions_path
